% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{kd.node}
\alias{kd.node}
\title{kd.node}
\description{
kd.node

kd.node
}
\examples{
root <- kd.node$new(1, 2, 3, 4)
note <- root$add_child(kd.node$new(2, 2, 5, 6))
note
root
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{item}}{node item}

\item{\code{axis}}{dimentsion to split}

\item{\code{left}}{left node}

\item{\code{right}}{right node}

\item{\code{parent}}{parent node}

\item{\code{location}}{location of node in the kd tree}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-kd.node-node_type}{\code{kd.node$node_type()}}
\item \href{#method-kd.node-set_location}{\code{kd.node$set_location()}}
\item \href{#method-kd.node-new}{\code{kd.node$new()}}
\item \href{#method-kd.node-add_child}{\code{kd.node$add_child()}}
\item \href{#method-kd.node-clone}{\code{kd.node$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kd.node-node_type"></a>}}
\if{latex}{\out{\hypertarget{method-kd.node-node_type}{}}}
\subsection{Method \code{node_type()}}{
correct the location for each node


the type of node, 0 represents root node; 1 represents left node; 2 represents right node
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kd.node$node_type()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{node}}{`kd.note`}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kd.node-set_location"></a>}}
\if{latex}{\out{\hypertarget{method-kd.node-set_location}{}}}
\subsection{Method \code{set_location()}}{
get and set the location of node by its item and parent's item
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kd.node$set_location()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kd.node-new"></a>}}
\if{latex}{\out{\hypertarget{method-kd.node-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kd.node$new(item = NULL, axis = NULL, left = NULL, right = NULL, parent = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{item}}{node item}

\item{\code{axis}}{dimension to split}

\item{\code{left}}{left node}

\item{\code{right}}{right node}

\item{\code{parent}}{parent node}

\item{\code{location}}{location of node in the kd tree}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kd.node-add_child"></a>}}
\if{latex}{\out{\hypertarget{method-kd.node-add_child}{}}}
\subsection{Method \code{add_child()}}{
add child node
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kd.node$add_child(child)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{child}}{`kd.note` to add}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kd.node-clone"></a>}}
\if{latex}{\out{\hypertarget{method-kd.node-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kd.node$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
