% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{kd.tree}
\alias{kd.tree}
\title{kd.tree}
\description{
kd.tree

kd.tree
}
\examples{
x <- c(2, 5, 9, 4, 8, 7)
y <- c(3, 4, 6, 7, 1, 2)
tree <- kd.tree$new(matrix(c(x, y ), ncol = 2))
tree$create()
tree
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{df}}{contain position of each node in `kd.tree`}

\item{\code{nodes}}{store the `kd.note` information}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-kd.tree-new}{\code{kd.tree$new()}}
\item \href{#method-kd.tree-create}{\code{kd.tree$create()}}
\item \href{#method-kd.tree-print}{\code{kd.tree$print()}}
\item \href{#method-kd.tree-clone}{\code{kd.tree$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kd.tree-new"></a>}}
\if{latex}{\out{\hypertarget{method-kd.tree-new}{}}}
\subsection{Method \code{new()}}{
transfer numeric vector to `kd.note`


get node type, 0 represents this node has no child nodes; 1 represents it only has left child; 2 represents it only has right child node; 3 represents has left and right children


transfer each children nodes to `kd.note`


get the nodes position in the tree
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kd.tree$new(X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{matrix object, each row is point, each column is the dimentsion}

\item{\code{axis}}{dimension to split}

\item{\code{ps}}{which rows need to split in input matrix, it should be a integer vector}

\item{\code{parent}}{parent node}

\item{\code{node}}{get the `kd.note` location}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kd.tree-create"></a>}}
\if{latex}{\out{\hypertarget{method-kd.tree-create}{}}}
\subsection{Method \code{create()}}{
create the `kd.note` to store in `nodes` and get the location of each node to store in `df`
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kd.tree$create()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kd.tree-print"></a>}}
\if{latex}{\out{\hypertarget{method-kd.tree-print}{}}}
\subsection{Method \code{print()}}{
print the nodes location in `kd.tree`
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kd.tree$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kd.tree-clone"></a>}}
\if{latex}{\out{\hypertarget{method-kd.tree-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kd.tree$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
