% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perceptron.R
\name{Perceptron}
\alias{Perceptron}
\title{Perceptron}
\description{
Perceptron
}
\examples{
#===============Example One===============
X <- iris[1:100,1:2]
y <- iris$Species[1:100]
m <- Perceptron$new(max_iteration = 10000)
m$fit(X, y)
m$plot()
#===============Example Two ===============
X <- matrix(c(3, 3, 4, 3, 1, 1), ncol = 2, byrow = TRUE)
y <- c(1, 1, -1)
fit <- Perceptron$new()
fit$fit(X, y, w = c(0,0), b = 0)
fit$plot()
#===============Example Three ===============
X <- iris[1:100,1:2]
y <- iris$Species[1:100]
set.seed(123)
idx <- sample(100, 50, replace = FALSE)
m <- Perceptron$new()
m$fit(X[idx, ], y[idx])
m$plot()
pre <- m$predict(X[-idx,], y[-idx])
m$plot(X[-idx, ], y[-idx])
#===============Example Four ===============
X <- iris[1:100,1:4]
y <- iris$Species[1:100]
set.seed(123)
idx <- sample(100, 50, replace = FALSE)
m <- Perceptron$new()
m$fit(X[idx, ], y[idx], seed = 888)
pre <- m$predict(X[-idx,], y[-idx])
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{X}}{is the depandent variables matrix}

\item{\code{y}}{is the response variable vector}

\item{\code{f_n}}{number of features}

\item{\code{s_n}}{number of samples}

\item{\code{w}}{weight vector}

\item{\code{b}}{intercept}

\item{\code{model}}{the final model}

\item{\code{labels}}{-1 and 1 stand for}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Perceptron-new}{\code{Perceptron$new()}}
\item \href{#method-Perceptron-fit}{\code{Perceptron$fit()}}
\item \href{#method-Perceptron-plot}{\code{Perceptron$plot()}}
\item \href{#method-Perceptron-predict}{\code{Perceptron$predict()}}
\item \href{#method-Perceptron-clone}{\code{Perceptron$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Perceptron-new"></a>}}
\if{latex}{\out{\hypertarget{method-Perceptron-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Perceptron$new(learn_ratio = 0.1, max_iteration = 2000, verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{learn_ratio}}{learn ratio, the default value is 0.1}

\item{\code{max_iteration}}{the max iteration}

\item{\code{verbose}}{show more messages}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Perceptron-fit"></a>}}
\if{latex}{\out{\hypertarget{method-Perceptron-fit}{}}}
\subsection{Method \code{fit()}}{
fit the data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Perceptron$fit(X, y, seed = 123, w = NULL, b = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{matrix contains dependent variable}

\item{\code{y}}{vector the response variable}

\item{\code{seed}}{used to `set.seed` for get random initial weights and intercept}

\item{\code{w}}{initial weights, `fit` will random initial a weights vector if `w` is NULL}

\item{\code{b}}{initial intercept, `fit` will random initial a value if `b` is NULL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Perceptron-plot"></a>}}
\if{latex}{\out{\hypertarget{method-Perceptron-plot}{}}}
\subsection{Method \code{plot()}}{
plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Perceptron$plot(X = NULL, y = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{same as the `X` of `fit`}

\item{\code{y}}{same as the `y` of `fit`}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Perceptron-predict"></a>}}
\if{latex}{\out{\hypertarget{method-Perceptron-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Perceptron$predict(new_X, new_y = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_X}}{the new data need to predict}

\item{\code{new_y}}{the y value of new data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Perceptron-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Perceptron-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Perceptron$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
