% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{queue}
\alias{queue}
\title{queue}
\description{
queue

queue
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{items}}{content of queue}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-queue-pop}{\code{queue$pop()}}
\item \href{#method-queue-is_empty}{\code{queue$is_empty()}}
\item \href{#method-queue-append}{\code{queue$append()}}
\item \href{#method-queue-print}{\code{queue$print()}}
\item \href{#method-queue-clone}{\code{queue$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-queue-pop"></a>}}
\if{latex}{\out{\hypertarget{method-queue-pop}{}}}
\subsection{Method \code{pop()}}{
pop
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{queue$pop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-queue-is_empty"></a>}}
\if{latex}{\out{\hypertarget{method-queue-is_empty}{}}}
\subsection{Method \code{is_empty()}}{
whether the queue empty
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{queue$is_empty()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-queue-append"></a>}}
\if{latex}{\out{\hypertarget{method-queue-append}{}}}
\subsection{Method \code{append()}}{
add value to queue
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{queue$append(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{value to add}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-queue-print"></a>}}
\if{latex}{\out{\hypertarget{method-queue-print}{}}}
\subsection{Method \code{print()}}{
print information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{queue$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-queue-clone"></a>}}
\if{latex}{\out{\hypertarget{method-queue-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{queue$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
