% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{kd.tree}
\alias{kd.tree}
\title{kd.tree}
\description{
kd.tree

kd.tree
}
\examples{
======================Example One====================
x <- c(2, 5, 9, 4, 8, 7)
y <- c(3, 4, 6, 7, 1, 2)
X <- matrix(c(x, y ), ncol = 2)
tree <- kd.tree$new(X)
tree
tree$plot()
tree$plot(T)
======================Example Two====================
X <- iris[1:10, 1:4]
tree <- kd.tree$new(X)
tree
tree$plot()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{df}}{the raw data to constructe `kd.tree`}

\item{\code{root}}{the root node}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-kd.tree-new}{\code{kd.tree$new()}}
\item \href{#method-kd.tree-get_graph}{\code{kd.tree$get_graph()}}
\item \href{#method-kd.tree-plot}{\code{kd.tree$plot()}}
\item \href{#method-kd.tree-print}{\code{kd.tree$print()}}
\item \href{#method-kd.tree-clone}{\code{kd.tree$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kd.tree-new"></a>}}
\if{latex}{\out{\hypertarget{method-kd.tree-new}{}}}
\subsection{Method \code{new()}}{
create the `kd.tree`
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kd.tree$new(X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{matrix-like data to construct `kd.tree`}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kd.tree-get_graph"></a>}}
\if{latex}{\out{\hypertarget{method-kd.tree-get_graph}{}}}
\subsection{Method \code{get_graph()}}{
return the igraph of the `kd.tree`
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kd.tree$get_graph()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kd.tree-plot"></a>}}
\if{latex}{\out{\hypertarget{method-kd.tree-plot}{}}}
\subsection{Method \code{plot()}}{
plot the tree
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kd.tree$plot(show_id = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{show_id}}{using the row number or feature content of raw data `X` as node label}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kd.tree-print"></a>}}
\if{latex}{\out{\hypertarget{method-kd.tree-print}{}}}
\subsection{Method \code{print()}}{
print the igraph
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kd.tree$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kd.tree-clone"></a>}}
\if{latex}{\out{\hypertarget{method-kd.tree-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{kd.tree$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
